pyproject.toml
[tool.poetry]
name = "hsa-update-server"
version = "0.1.0"
description = ""
authors = ["Anton Shusharin <antonshusharin01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-debug-toolbar = "^4.2.0"
gunicorn = "^21.2.0"
redis = "^5.0.1"
hiredis = "^2.3.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
